set up authentication and authorization for deleting a post

+++ controllers/api/v1/posts_api.js
@@ -27,14 +27,22 @@ module.exports.destroy=async function(req, res){
     try{
         let post=await Post.findById(req.params.id);
 
-            post.remove();
+            if(post.user==req.user.id){
 
-            await Comment.deleteMany({post : req.params.id});
+                post.remove();
+
+                await Comment.deleteMany({post : req.params.id});
 
 
-            return res.status(200).json({
-                message : 'Post and associated comments deleted!'
-            });
+                return res.status(200).json({
+                    message : 'Post and associated comments deleted!'
+                });
+            }
+            else{
+                return res.status(401).json({
+                    message : 'You cannot delete this post!'
+                });
+            }
     }catch(err){
         return res.status(500).json({
             message : 'Internal Server Error'
